moved remove action on curation list
